+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Changelog
 =========
 
+ * Added Laravel 5.6, 5.7 and 5.8 support
  * Dropped Symfony 2.8 support
  * Dropped PHP 7.0 support